[run ci] address cmments
