+++ BodoSQL/bodosql/tests/test_types/test_s3_tables_iceberg_catalog.py
@@ -99,11 +99,15 @@ def test_s3_tables_catalog_iceberg_write(s3_tables_catalog, memory_leak_check):
         exception_occurred_in_test_body = True
         raise e
     finally:
-        catalog = S3TablesCatalog(
-            "s3_tables_catalog", **{"s3tables.warehouse": s3_tables_catalog.warehouse}
-        )
         try:
-            catalog.drop_table(f"write_namespace.{table_name}")
+            run_rank0(
+                lambda: (
+                    S3TablesCatalog(
+                        "s3_tables_catalog",
+                        **{"s3tables.warehouse": s3_tables_catalog.warehouse},
+                    ).drop_table(f"write_namespace.{table_name}")
+                )
+            )()
         except Exception:
             if exception_occurred_in_test_body:
                 pass
