+++ bodo/io/iceberg/stream_iceberg_write.py
@@ -5,12 +5,13 @@ Not intended for external use
 
 from __future__ import annotations
 
+import importlib
+import importlib.util
 import operator
 import os
 import traceback
 import typing as pt
 
-import llvmlite.binding as ll
 import numba
 import pandas as pd
 import pyarrow as pa
@@ -43,9 +44,13 @@ from bodo.io.helpers import (
 from bodo.io.iceberg.catalog import conn_str_to_catalog
 from bodo.io.iceberg.common import IcebergConnectionType
 from bodo.io.iceberg.theta import (
+    _write_puffin_file,
     commit_statistics_file,
     fetch_puffin_metadata,
+    get_default_theta_sketch_columns,
     get_old_statistics_file_path,
+    get_supported_theta_sketch_columns,
+    init_theta_sketches_wrapper,
     table_columns_enabled_theta_sketches,
     table_columns_have_theta_sketches,
     theta_sketch_collection_type,
@@ -60,12 +65,8 @@ from bodo.io.iceberg.write import (
     transaction_type,
     wrap_start_write,
 )
-from bodo.libs import puffin_file, theta_sketches
 from bodo.libs.array import (
-    array_info_type,
     array_to_info,
-    delete_info,
-    info_to_array,
     py_table_to_cpp_table,
 )
 from bodo.libs.bool_arr_ext import alloc_false_bool_array
@@ -90,41 +91,7 @@ if pt.TYPE_CHECKING:  # pragma: no cover
     from pyiceberg.table import Transaction
 
 
-ll.add_symbol("init_theta_sketches", theta_sketches.init_theta_sketches_py_entrypt)
-ll.add_symbol("delete_theta_sketches", theta_sketches.delete_theta_sketches_py_entrypt)
-ll.add_symbol(
-    "fetch_ndv_approximations", theta_sketches.fetch_ndv_approximations_py_entrypt
-)
-ll.add_symbol("write_puffin_file", puffin_file.write_puffin_file_py_entrypt)
-ll.add_symbol("read_puffin_file_ndvs", puffin_file.read_puffin_file_ndvs_py_entrypt)
-ll.add_symbol(
-    "get_supported_theta_sketch_columns",
-    theta_sketches.get_supported_theta_sketch_columns_py_entrypt,
-)
-ll.add_symbol(
-    "get_default_theta_sketch_columns",
-    theta_sketches.get_default_theta_sketch_columns_py_entrypt,
-)
-
 ICEBERG_WRITE_PARQUET_CHUNK_SIZE = int(256e6)
-
-
-def get_env_value(env_var):  # pragma: no cover
-    pass
-
-
-@overload(get_env_value)
-def overload_get_env_value(env_var):
-    """
-    Returns the current runtime value of an environment variable
-    """
-
-    def impl(env_var):  # pragma: no cover
-        with bodo.no_warning_objmode(env_value="string"):
-            env_value = os.environ.get(env_var, "0")
-        return env_value
-
-    return impl
 
 
 def get_enable_theta():  # pragma: no cover
@@ -139,316 +106,19 @@ def overload_get_enable_theta():
 
     def impl():  # pragma: no cover
         with bodo.no_warning_objmode(ret_var="bool_"):
-            ret_var = bodo.enable_theta_sketches
+            if not bodo.enable_theta_sketches:
+                return False
+
+            try:
+                if importlib.util.find_spec("bodo_iceberg_connector") is None:
+                    return False
+            except ValueError:
+                return False
+
+            ret_var = True
         return ret_var
 
     return impl
-
-
-@intrinsic
-def _init_theta_sketches(
-    typingctx,
-    column_bitmask_t,
-):
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),  # table_info*
-            [
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="init_theta_sketches"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        theta_sketch_collection_type(array_info_type),
-        codegen,
-    )
-
-
-def get_supported_theta_sketch_columns(iceberg_pyarrow_schema):  # pragma: no cover
-    pass
-
-
-@overload(get_supported_theta_sketch_columns)
-def overload_get_supported_theta_sketch_columns(iceberg_pyarrow_schema):
-    """
-    Returns a boolean array indicating which columns have types that can
-    support theta sketches.
-    """
-    arr_type = bodo.boolean_array_type
-
-    def impl(iceberg_pyarrow_schema):  # pragma: no cover
-        res_info = _get_supported_theta_sketch_columns(iceberg_pyarrow_schema)
-        res = info_to_array(res_info, arr_type)
-        delete_info(res_info)
-        return res
-
-    return impl
-
-
-@intrinsic
-def _get_supported_theta_sketch_columns(typingctx, iceberg_pyarrow_schema_t):
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),  # array_info*
-            [
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="get_supported_theta_sketch_columns"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        array_info_type(pyarrow_schema_type),
-        codegen,
-    )
-
-
-def get_default_theta_sketch_columns(iceberg_pyarrow_schema):  # pragma: no cover
-    pass
-
-
-@overload(get_default_theta_sketch_columns)
-def overload_get_default_theta_sketch_columns(iceberg_pyarrow_schema):
-    """
-    Returns a boolean array indicating which columns have types that output
-    theta sketches by default.
-    """
-    arr_type = bodo.boolean_array_type
-
-    def impl(iceberg_pyarrow_schema):  # pragma: no cover
-        res_info = _get_default_theta_sketch_columns(iceberg_pyarrow_schema)
-        res = info_to_array(res_info, arr_type)
-        delete_info(res_info)
-        return res
-
-    return impl
-
-
-@intrinsic
-def _get_default_theta_sketch_columns(typingctx, iceberg_pyarrow_schema_t):
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),  # array_info*
-            [
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="get_default_theta_sketch_columns"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        array_info_type(pyarrow_schema_type),
-        codegen,
-    )
-
-
-def init_theta_sketches_wrapper(column_bitmask):  # pragma: no cover
-    pass
-
-
-@overload(init_theta_sketches_wrapper)
-def overload_init_theta_sketches_wrapper(column_bit_mask):
-    """
-    Creates a new theta sketch collection when starting to write an Iceberg table.
-
-    Args:
-        column_bit_mask (Boolean Array): An array of booleans indicating which columns
-            have theta sketches enabled.
-
-    Returns:
-        C++ Object: A new theta sketch collection object, which is effectively a pointer
-            to an array of theta sketches, with null entries for columns without sketches.
-    """
-
-    def impl(column_bit_mask):  # pragma: no cover
-        return _init_theta_sketches(array_to_info(column_bit_mask))
-
-    return impl
-
-
-@intrinsic
-def _iceberg_writer_fetch_theta(typingctx, array_info_t):
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),  # array_info*
-            [
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="fetch_ndv_approximations"
-        )
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        array_info_type(theta_sketch_collection_type),
-        codegen,
-    )
-
-
-def iceberg_writer_fetch_theta(writer):
-    pass
-
-
-@overload(iceberg_writer_fetch_theta)
-def overload_iceberg_writer_fetch_theta(writer):
-    """
-    Fetches the current values of the theta sketch approximations
-    of NDV for each column in an iceberg writer. For each column
-    that does not have a theta sketch, returns null instead. Largely
-    used for testing purposes.
-    """
-    arr_type = bodo.FloatingArrayType(types.float64)
-
-    def impl(writer):  # pragma: no cover
-        res_info = _iceberg_writer_fetch_theta(writer["theta_sketches"])
-        res = info_to_array(res_info, arr_type)
-        delete_info(res_info)
-        return res
-
-    return impl
-
-
-@intrinsic
-def _read_puffin_file_ndvs(typingctx, puffin_loc_t, bucket_region_t, iceberg_schema_t):
-    def codegen(context, builder, sig, args):
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),  # array_info*
-            [
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-                lir.IntType(8).as_pointer(),
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="read_puffin_file_ndvs"
-        )
-
-        ret = builder.call(fn_tp, args)
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        return ret
-
-    return (
-        array_info_type(types.voidptr, types.voidptr, pyarrow_schema_type),
-        codegen,
-    )
-
-
-def read_puffin_file_ndvs(puffin_file_loc):  # pragma: no cover
-    pass
-
-
-@overload(read_puffin_file_ndvs)
-def overload_read_puffin_file_ndvs(puffin_file_loc, iceberg_schema):
-    """
-    Reads the NDV values from a puffin file. This is used for testing purposes
-    to verify that the theta sketches are being written correctly.
-    """
-    arr_type = bodo.FloatingArrayType(types.float64)
-
-    def impl(puffin_file_loc, iceberg_schema):  # pragma: no cover
-        bucket_region = bodo.io.fs_io.get_s3_bucket_region_wrapper(
-            puffin_file_loc, parallel=True
-        )
-        res_info = _read_puffin_file_ndvs(
-            unicode_to_utf8(puffin_file_loc),
-            unicode_to_utf8(bucket_region),
-            iceberg_schema,
-        )
-        res = info_to_array(res_info, arr_type)
-        delete_info(res_info)
-        return res
-
-    return impl
-
-
-@intrinsic
-def _write_puffin_file(
-    typingctx,
-    puffin_file_loc_t,
-    bucket_region_t,
-    snapshot_id_t,
-    sequence_number_t,
-    theta_sketches_t,
-    output_pyarrow_schema_t,
-    arrow_fs_t,
-    exist_puffin_loc_t,
-):
-    def codegen(context, builder, sig, args):
-        (
-            puffin_file_loc,
-            bucket_region,
-            snapshot_id,
-            sequence_number,
-            theta_sketches,
-            output_pyarrow_schema,
-            arrow_fs,
-            exist_puffin_loc,
-        ) = args
-        fnty = lir.FunctionType(
-            lir.IntType(8).as_pointer(),
-            [
-                lir.IntType(8).as_pointer(),  # puffin_file_loc
-                lir.IntType(8).as_pointer(),  # bucket_region
-                lir.IntType(64),  # snapshot_id
-                lir.IntType(64),  # sequence_number
-                lir.IntType(8).as_pointer(),  # theta_sketches
-                lir.IntType(8).as_pointer(),  # output_pyarrow_schema
-                lir.IntType(8).as_pointer(),  # arrow_fs
-                lir.IntType(8).as_pointer(),  # exist_puffin_loc
-            ],
-        )
-        fn_tp = cgutils.get_or_insert_function(
-            builder.module, fnty, name="write_puffin_file"
-        )
-        ret = builder.call(
-            fn_tp,
-            [
-                puffin_file_loc,
-                bucket_region,
-                snapshot_id,
-                sequence_number,
-                theta_sketches,
-                output_pyarrow_schema,
-                arrow_fs,
-                exist_puffin_loc,
-            ],
-        )
-        bodo.utils.utils.inlined_check_and_propagate_cpp_exception(context, builder)
-        # Wrap the PyObject* in a version that can track reference counts.
-        return bodo.utils.py_objs.create_struct_from_pyobject(
-            sig.return_type, ret, context, builder, context.get_python_api(builder)
-        )
-
-    return (
-        types.statistics_file_type(
-            types.voidptr,  # Pass UTF-8 string as void*
-            types.voidptr,  # const Pass UTF-8 string as void*
-            types.int64,
-            types.int64,
-            theta_sketch_collection_type,
-            output_pyarrow_schema_t,
-            pyarrow_fs_type,
-            types.voidptr,  # Pass UTF-8 string as void*
-        ),
-        codegen,
-    )
 
 
 class IcebergWriterType(StreamingStateType):
@@ -484,6 +154,10 @@ iceberg_writer_payload_type = IcebergWriterPayloadType()
 
 
 iceberg_writer_payload_members = (
+    # Connection String to Iceberg Table
+    ("conn", types.unicode_type),
+    # Table identifier to write to
+    ("table_id", types.unicode_type),
     # Action Iceberg takes if table exists. Currently one of "fail", "append", "replace"
     ("if_exists", types.unicode_type),
     # Location of the data/ folder in the warehouse
@@ -893,6 +567,8 @@ def gen_iceberg_writer_init_impl(
 
         # Initialize writer
         writer = iceberg_writer_alloc(iceberg_writer_type)
+        writer["conn"] = conn
+        writer["table_id"] = table_id
         writer["if_exists"] = if_exists
         writer["table_loc"] = table_loc
         writer["iceberg_schema_str"] = iceberg_schema_str
@@ -1226,8 +902,10 @@ def gen_iceberg_writer_append_table_impl_inner(
                     writer["arrow_fs"],
                     unicode_to_utf8(old_puffin_file_path),
                 )
+                conn_str = writer["conn"]
+                table_id = writer["table_id"]
                 with bodo.no_warning_objmode():
-                    commit_statistics_file(txn, statistic_file_info)
+                    commit_statistics_file(conn_str, table_id, statistic_file_info)
 
             # Delete the theta sketches. An object exists even if there are no sketches.
             delete_theta_sketches(writer["theta_sketches"])
