+++ bodo/tests/test_iceberg/test_iceberg_theta.py
@@ -9,9 +9,11 @@ import bodo
 from bodo.io.iceberg.stream_iceberg_write import (
     iceberg_writer_append_table,
     iceberg_writer_init,
-    read_puffin_file_ndvs,
 )
-from bodo.io.iceberg.theta import table_columns_have_theta_sketches
+from bodo.io.iceberg.theta import (
+    read_puffin_file_ndvs,
+    table_columns_have_theta_sketches,
+)
 from bodo.ir.sql_ext import remove_iceberg_prefix
 from bodo.tests.iceberg_database_helpers.metadata_utils import (
     get_metadata_field,
