+++ bodo/transforms/distributed_analysis.py
@@ -1376,7 +1376,7 @@ class DistributedAnalysis:
 
         if fdef == (
             "iceberg_writer_fetch_theta",
-            "bodo.io.iceberg.stream_iceberg_write",
+            "bodo.io.iceberg.theta",
         ):
             # Used to obtain the current value of a theta sketch collection from
             # an Iceberg writer as an array, where each row is the current estimate
@@ -1385,7 +1385,7 @@ class DistributedAnalysis:
             _set_REP(self.typemap, self.metadata, self.diag_info, lhs, array_dists)
             return
 
-        if fdef == ("read_puffin_file_ndvs", "bodo.io.iceberg.stream_iceberg_write"):
+        if fdef == ("read_puffin_file_ndvs", "bodo.io.iceberg.theta"):
             # Used to the ndvs from a puffin file for testing.
             _set_REP(self.typemap, self.metadata, self.diag_info, lhs, array_dists)
             return
