+++ iceberg/bodo_iceberg_connector/puffin.py
@@ -46,7 +46,9 @@ class StatisticsFile:
         return StatisticsFile(-1, "", -1, -1, [])
 
 
-def table_columns_have_theta_sketches(conn_str: str, db_name: str, table_name: str):
+def table_columns_have_theta_sketches(
+    conn_str: str, table_id: str
+) -> pd.arrays.BooleanArray:
     """
     Determine which columns in a given table have theta sketches. The returned
     result is a boolean array where the ith element is True if the ith column
@@ -64,40 +66,15 @@ def table_columns_have_theta_sketches(conn_str: str, db_name: str, table_name: s
     """
     catalog_type, _ = parse_conn_str(conn_str)
     handler = get_catalog(conn_str, catalog_type)
-    hasSketches = handler.tableColumnsHaveThetaSketches(db_name, table_name)
-    return pd.array(hasSketches, dtype="boolean")
+    hasSketches = handler.tableColumnsHaveThetaSketches(table_id)
+    return pd.array(hasSketches, dtype="boolean")  # type: ignore[return]
 
 
-def table_columns_enabled_theta_sketches(conn_str: str, db_name: str, table_name: str):
-    """
-    Determine which columns in a given table have theta sketches ENABLED. The returned
-    result is a boolean array where the ith element is True if the ith column
-    has a theta sketch enabled via the table property, and False otherwise. The indices are based on current
-    column locations in the table schema, not field id.
-    See implementation in BodoIcebergHandler.java for more details.
-
-    Args:
-        conn_str (str): The iceberg connection string.
-        db_name (str): The table's database name.
-        table_name (str): The table name.
-
-    Returns:
-        pd.array[bool]: A boolean array where the ith element is True if the ith
-        column has a theta sketch enabled, and False otherwise.
-    """
-    catalog_type, _ = parse_conn_str(conn_str)
-    handler = get_catalog(conn_str, catalog_type)
-    enabledSketches = handler.tableColumnsEnabledThetaSketches(db_name, table_name)
-    return pd.array(enabledSketches, dtype="boolean")
-
-
-def get_old_statistics_file_path(
-    txn_id: int, conn_str: str, db_name: str, table_name: str
-):
+def get_old_statistics_file_path(conn_str: str, table_id: str) -> str:
     """
     Get the old puffin file path from the connector. We know that the puffin file
     must exist because of previous checks.
     """
     catalog_type, _ = parse_conn_str(conn_str)
     handler = get_catalog(conn_str, catalog_type)
-    return normalize_loc(handler.getStatisticsFilePath(txn_id))
+    return normalize_loc(handler.getStatisticsFilePath(table_id))
